fixing zsh and terminal profiles

+++ scripts/Setup-Ubuntu.ps1
@@ -155,28 +155,36 @@ if ($LASTEXITCODE -ne 0) { throw "failed to install gh cli." }
 if ($LASTEXITCODE -ne 0) { throw "failed to install zsh." }
 
 # fzf
-& $distro run "rm -rf ~/.fzf"
-if ($LASTEXITCODE -ne 0) { throw "failed to clean up old fzf installation." }
-& $distro run "git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf"
-if ($LASTEXITCODE -ne 0) { throw "failed to clone fzf." }
-& $distro run "~/.fzf/install --all"
-if ($LASTEXITCODE -ne 0) { throw "failed to install fzf." }
+$foundFzf = & $distro run "source ~/.zshrc &> /dev/null; type fzf"
+if ($foundFzf -match "fzf not found") {
+    # clean up and install oh my zsh
+    & $distro run "rm -rf ~/.fzf"
+    if ($LASTEXITCODE -ne 0) { throw "failed to clean up old fzf installation." }
+    & $distro run "git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf"
+    if ($LASTEXITCODE -ne 0) { throw "failed to clone fzf." }
+    & $distro run "~/.fzf/install --all"
+    if ($LASTEXITCODE -ne 0) { throw "failed to install fzf." }
+}
 
 # oh-my-zsh
-& $distro run "sh -c 'curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh | sh'"
-if ($LASTEXITCODE -ne 0) { throw "failed to install oh my zsh." }
+$foundOmz = & $distro run "source ~/.zshrc &> /dev/null; type omz"
+if ($foundOmz -match "omz not found") {
+    # install oh my zsh
+    & $distro run "sh -c 'curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh | sh'"
+    if ($LASTEXITCODE -ne 0) { throw "failed to install oh my zsh." }
+
+    # oh-my-zsh plugins
+    & $distro run "git clone https://github.com/zsh-users/zsh-autosuggestions ~/.oh-my-zsh/custom/plugins/zsh-autosuggestions"
+    if ($LASTEXITCODE -ne 0) { throw "failed to install zsh-autosuggestions plugin." }
+    & $distro run "git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ~/.oh-my-zsh/custom/plugins/zsh-syntax-highlighting"
+    if ($LASTEXITCODE -ne 0) { throw "failed to install zsh-syntax-highlighting plugin." }
+}
 
 # profile settings for zshrc and fzf
 & $distro run "curl -fsSL https://raw.githubusercontent.com/predragstefanovic/winplay/refs/heads/main/config/zsh/.zshrc -o ~/.zshrc"
 if ($LASTEXITCODE -ne 0) { throw "failed to install oh my zsh profile settings." }
 & $distro run "curl -fsSL https://raw.githubusercontent.com/predragstefanovic/winplay/refs/heads/main/config/zsh/.fzf.zsh -o ~/.fzf.zsh"
 if ($LASTEXITCODE -ne 0) { throw "failed to install oh my fzf settings." }
-
-# oh-my-zsh plugins
-& $distro run "git clone https://github.com/zsh-users/zsh-autosuggestions ~/.oh-my-zsh/custom/plugins/zsh-autosuggestions"
-if ($LASTEXITCODE -ne 0) { throw "failed to install zsh-autosuggestions plugin." }
-& $distro run "git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ~/.oh-my-zsh/custom/plugins/zsh-syntax-highlighting"
-if ($LASTEXITCODE -ne 0) { throw "failed to install zsh-syntax-highlighting plugin." }
 
 # activate zsh to be the default shell
 & $distro run "sudo -S <<< ${password} chsh -s /usr/bin/zsh $username"
